Fixing require problem on /src/util.php
